Conflicts:
	ZottaOS/CORTEX-Mx/ZottaOS_CortexMx.h

+++ ZottaOS/ZottaOSHardPA.h
@@ -388,20 +388,20 @@ void OSScheduleSuspendedTask(void *event);
 ** er to an interrupt service routine that is called when an interrupt from the device
 ** and for the particular source should be handled.
 ** Parameters:
-**   (1) (UINT8) entry: entry to the internal ISR interrupt vector;
+**   (1) (UINT16) entry: entry to the internal ISR interrupt vector;
 **   (2) (void *) descriptor: ISR descriptor for the interrupt. */
-void OSSetISRDescriptor(UINT8 entry, void *descriptor);
+void OSSetISRDescriptor(UINT16 entry, void *descriptor);
 
 /* OSGetISRDescriptor: Returns the ISR descriptor associated with the internal table of
 ** the kernel. This function is made available to application tasks so that the inserted
 ** ISR descriptor can be returned to the user in order to read or to store his applica-
 ** tion data.
-** Parameter: (UINT8) Index in the ISR device interrupt vector where the ISR descriptor
+** Parameter: (UINT16) Index in the ISR device interrupt vector where the ISR descriptor
 **    is held.
 ** Returned value: (void *) The requested ISR descriptor is returned. If no previous call
 **    to OSSetISRDescriptor was previously made for the specified entry, the returned
 **    value is undefined. */
-void *OSGetISRDescriptor(UINT8 entry);
+void *OSGetISRDescriptor(UINT16 entry);
 
 
 /* INTER-TASK COMMUNICATION AND I/O BUFFERS FOR INTERRUPT HANDLERS ------------------- */
